pyproject.toml
[project]
name = "zabbix-erp-sincronismo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "vieutil>=0.5",
]
